% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildStructandSeries.R
\name{adjustRgroup}
\alias{adjustRgroup}
\title{Add extra atoms and unsaturations before or after an R group}
\usage{
adjustRgroup(genSmiles,R_format,ExtraAtoms=NULL,RDB=NULL,R_range=NULL,atStart=TRUE)
}
\arguments{
\item{genSmiles}{A "Generic" \code{SMILES} code to expand, with R groups encoded in square brackets.
The R group to replace is defined in R_format.}

\item{R_format}{The format of the R group to replace, including square brackets, e.g.
\code{[R], [R1], [R2]}.}

\item{ExtraAtoms}{A valid \code{SMILES} code representing the single unit to add.
Single-valent atoms should be included with their attached atom. Examples
\code{"C(=O)", "C(Br)", "C(O)", "O"}. Must be valid \code{SMILES} to generate valid structures.
Use with caution - it is highly recommended to view the structures with
\code{\link{renderSMILES.CDKdepict}} or \code{\link{renderSMILES.rcdk}} to detect unwanted
behaviour.}

\item{RDB}{A number (Ring and Double Bond count) indicating the degree of unsaturations to add.
This number determines how many \code{"C=C"} units are added. Note that no double-bond stereoisomerism
is added. If \code{R_range} is given, this will be adjusted according to the number of \code{"C=C"} groups
added, unless a negative range would be generated, in this case the original range is returned.}

\item{R_range}{The range for the R group (start-end). Will be corrected by \code{2*RDB} if given.}

\item{atStart}{Default \code{TRUE} adds extra atoms and unsaturated units before the R group,
\code{FALSE} adds afterwards. Triply substituted Cs should only be added after terminal Rs, for instance.}
}
\value{
Returns a vector containing the resulting SMILES codes and,
if range is given, the new min and max values for range.
}
\description{
This works behind the scenes of \code{buildSmiles} on individual R groups, but can be
used independently. This performs basic string manipulation on SMILES codes to add
additional atoms (as valid SMILES codes) and unsaturations before or after R groups
prior to generation of systematic homologues with RtoSMILES. This facilitates the generation
of identifiers for structures that may not exist in databases for use in MS workflows.
This requires a decent understanding of SMILES as the text manipulation can have
unexpected consequences.
It is highly recommeded to view these structures to ensure the groups are added as intended.
}
\examples{

test_genSmiles <- adjustRgroup("OC(=O)[R]","[R]",ExtraAtoms="C(=O)")
test_genSmiles <- adjustRgroup("OC(=O)[R]","[R]",ExtraAtoms="C(Br)")
test_genSmiles <- adjustRgroup("OC(=O)[R]","[R]",RDB=2,atStart=FALSE)
buildCDKdepictURL(test_genSmiles)
# if the range is returned, need to only depict first entry
test_genSmiles <- adjustRgroup("OC(=O)[R]","[R]",ExtraAtoms="C(=O)",RDB=2,atStart=TRUE, R_range = c(4,6))
test_genSmiles <- adjustRgroup("OC(=O)[R]","[R]",RDB=2,atStart=TRUE, R_range = c(2,20))
buildCDKdepictURL(test_genSmiles[1])

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{buildSmiles}}, \code{\link{splitRrange}}, \code{\link{RtoSMILES}},
To view created SMILES: \code{\link{renderSMILES.CDKdepict}}, \code{\link{renderSMILES.rcdk}}.
}

