% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildStructandSeries.R
\name{build.homol}
\alias{build.homol}
\title{Build molecular formulae for a homologue series}
\usage{
build.homol(start_formula, building_block, series_name, n_start, n_max, rt=15)
}
\arguments{
\item{start_formula}{Molecular formula to start building the series. Standardised with
\code{checkform}, along with \code{building_block}.}

\item{building_block}{Formula of the "building block" to add n times to \code{start_formula}.}

\item{series_name}{Codename for the series. Short names are recommended for downstream use.}

\item{n_start}{Starting number for member naming only. Formula building goes from n=1 (
\code{start_formula}) to n=(\code{n_max}-\code{n_start}).}

\item{n_max}{Maximum number to calculate.}

\item{rt}{Default \code{15} (minutes); can be adjusted for different systems and anticipated
series behaviour as desired. For first round screening, RT in the middle of the elution is
recommended, such that a symmetrical window over the whole program can be used.}
}
\value{
Returns a \code{n_max-n_start} by \code{4} array containing member names, formulas, RTs
and neutral monoisotopic mass.
}
\description{
Given a starting formula and a "building block", this builds a series for formulas
from \code{n_start} to \code{n_max}, adding a retention time for direct use in
suspect screening. Naming is built on \code{series_name} and \code{n}. Formula
manipulations are performed with \code{enviPat} and \code{RMassBank}.
Can be used if generic structures are not available for \code{\link{buildSmiles}}.
}
\examples{
las <- build.homol("C16H26O3S","CH2","LAS",10,14)
spacs <- build.homol("C9H10O5S","CH2","SPAC",3,15)
spadcs <- build.homol("C9H8O7S","CH2","SPADC",1,15)
c12aes <- build.homol("C14H30O5S","C2H4O","C12AES",1,10)
c13aes <- build.homol("C15H32O5S","C2H4O","C13AES",1,6)
npeos <- build.homol("C17H28O5S","C2H4O","NPEO_SO4_",1,8)

}
\seealso{
\code{\link{buildSmiles}}, \code{\link{check_chemform}}, \code{\link{mergeform}},
\code{\link{enviPat}}, \code{\link{findMz.formula}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
