% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructurePlotting.R
\name{buildCDKdepictURL}
\alias{buildCDKdepictURL}
\title{Build URL for CDK Depict}
\usage{
buildCDKdepictURL(smiles, style="bow",title=NULL, abbr="off", smarts=NULL,
suppressh=TRUE,showtitle=TRUE,zoom=2,annotate="none")
}
\arguments{
\item{smiles}{A valid \code{SMILES} code or a \code{CxSMILES} for advanced plotting.
The latter is not yet accessible in R and must be compiled manually or externally.}

\item{style}{The plotting style, one of 5 options \code{("cow","cob","bow","wob","nob")},
i.e. color on white, color on black, black on white, white on black, neon on black. The
appearance of \code{SMARTS} and code{annotate} options change with style.}

\item{title}{(optional) title to display under structure.}

\item{abbr}{Default \code{"off"}, this controls whether the structure is displayed "as is"
\code{"off"} or whether groups (\code{"groups"}), reagents (\code{"reagents"}) or both reagents
and groups (\code{"on"}) are abbreviated.}

\item{smarts}{(optional) \code{SMARTS} codes can be entered to highlight parts of the molecule.}

\item{suppressh}{Default \code{TRUE} suppresses the display of Hs. \code{FALSE} does the oppposite.}

\item{showtitle}{Default \code{TRUE} displays the title below the molecule. \code{FALSE} turns
this off.}

\item{zoom}{Controls the size of the image, default 2, increasing to 5 results in a large
structure in the web browser. Warnings at zoom>=10.}

\item{annotate}{(optional) Features to interpret the structure. Default \code{"none"}, other
options Atom Numbers, Atom Mapping, Color Map, Atom Value
(\code{"number", "mapidx", "colmap", "atomvalue"}) respectively.}
}
\value{
Returns a URL for direct use in a browser, or for rendering in R.
}
\description{
This builds a URL for the CDK Depict website for direct viewing or for
retrieving svgs for \code{renderSMILES.CDKdepict}. Contains all the options
in Version 0.2, reverting to default options if invalid entries are specified.
}
\examples{
buildCDKdepictURL("c1ccccc1")
buildCDKdepictURL("CN1C=NC2=C1C(=O)N(C(=O)N2C)C",style="cow",title="caffeine",
                  abbr="off",smarts="c1cncnc1",suppressh=FALSE,showtitle=TRUE,zoom=5,annotate="number")
buildCDKdepictURL("CCOCCOCCO",style="cow")
buildCDKdepictURL("CCOCCOCCO |Sg:n:3,4,5::ht|",style="cow", smarts="CCO")
buildCDKdepictURL("CCOCCOCCO |Sg:n:3,4,5::ht|",style="cow", smarts="CCO",title="PEGn")
buildCDKdepictURL("OS(=O)(=O)c1ccc(cc1)C([R1]C(=O)O)[R2]C(=O)O", style="cow",title="SPADCs",
abbr="off",suppressh=TRUE,showtitle=TRUE,zoom=5,annotate="number",smarts="C(=O)O")
buildCDKdepictURL("OS(=O)(=O)c1ccc(cc1)C(CC(=O)O)CC(=O)O |Sg:n:15:m:ht,Sg:n:11:n:ht|", style="cow",
title="SPADCs, n+m=0-5",abbr="off",suppressh=TRUE,showtitle=TRUE,zoom=5,annotate="number",smarts="C(=O)O")

}
\author{
Emma Schymanski (R wrapper, <emma.schymanski@uni.lu>), John Mayfield (CDK Depict)
}
\seealso{
\code{\link{renderSMILES.CDKdepict}} to use the resulting structure in plotting,
\code{\link{renderSMILES.rcdk}} for alternative rendering.
}

