% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompTox.R
\name{buildCompToxURL}
\alias{buildCompToxURL}
\title{Build URL for CompTox Chemistry Dashboard}
\usage{
buildCompToxURL(search_string, type="simple", add_utf8=FALSE, browse=FALSE,
mass_entry="", single_component=TRUE, no_isotopes=TRUE)
}
\arguments{
\item{search_string}{A string to search CompTox. Must be compatible with one of the \code{type}
entries to perform a valid search.}

\item{type}{Type of search to perform, one of \code{("simple", "name", "cas","DTXSID","InChIKey",
"skeleton", "similarity","mass_ppm", "mass_Da", "mass_range", "formula")}. See Details.}

\item{add_utf8}{If \code{TRUE}, adds utf8 character into the URL for compatability with IE8.
If this is added, it is preferable to copy/paste the link into the browser rather than use
\code{browse=TRUE} to avoid encoding issues.}

\item{browse}{Default \code{FALSE} returns just a URL for this function, if \code{TRUE}, the
URL is opened in the default browser as well.}

\item{mass_entry}{This must be non-empty for all mass searches; for \code{mass_ppm} it is the ppm
value to use, for \code{mass_Da} the absolute mass error and for \code{mass_range} it is the
higher mass entry in the range (and \code{search_string} must be the lower entry).}

\item{single_component}{Default \code{TRUE} means only single component records are searched;
\code{FALSE} includes mixtures. Only used in combination with mass and formula searches.}

\item{no_isotopes}{Default \code{TRUE} eliminates non-standard isotope species. \code{FALSE}
includes these. Used only in mass searches.}
}
\value{
Returns a URL for direct use in a browser, optionally opening it in the default browser
}
\description{
This builds a URL for searching the US EPA CompTox Chemistry Dashboard, optionally
sending this direct to the default browser.
}
\details{
Currently the CompTox Dashboard supports name, CAS, DTXSID, InChIKey and InChIKey skeleton
searches as a simple text-based search requiring only \code{search_string}.

The \code{type="simple"}
covers all these categories, alternatively the respective \code{type} can be used as follows:

name = \code{"name"}, CAS reference number = \code{"cas"},
DSSToxID (Dashboard identifier) = \code{"DTXSID"} (including the letters DTXSID!), InChIKey =
\code{InChIKey} and InChIKey skeleton (first block) search = \code{"skeleton"}.

The similarity search (\code{type="similarity"}) requires a DTXSID as \code{search_string}.
Mass-based searches are a combination of \code{search_string} and \code{mass_entry} (see above).

Formula searches require a molecular formula as \code{search_string} and type=\code{"formula"}.
Note that a SMILES search is not yet supported in the Dashboard.
}
\examples{
buildCompToxURL("benzene")
buildCompToxURL("nicotine", browse=TRUE)

# if you add utf8, better to copy and paste link into browswer rather than use browse=TRUE.
buildCompToxURL("1912-24-9", type = "cas", add_utf8 = TRUE)

buildCompToxURL("MXWJVTOOROXGIU-UHFFFAOYSA-N", browse=TRUE)
buildCompToxURL("SNICXCGAKADSCV", type="skeleton", browse=TRUE)

# this searches just for the exact ID
buildCompToxURL("DTXSID9020112", browse=TRUE)
# this searches for similar substances, appearing lower down in the record
buildCompToxURL("DTXSID9020112", type="similarity", browse=TRUE)

# Molecular formula search: (default with single component)
buildCompToxURL("C8H14ClN5", type="formula", browse=TRUE)
buildCompToxURL("C8H14ClN5", type="formula", browse=TRUE, single_component = FALSE)
# These two links return the same result, this shouldn't be the case.
# Single component part does not seem to be working (also in web interface). Comment submitted.
buildCompToxURL("C8H15Cl2N5", type="formula", browse=TRUE, single_component = TRUE)
buildCompToxURL("C8H15Cl2N5", type="formula", browse=TRUE, single_component = FALSE)

# Mass search with ppm and Da
buildCompToxURL("162.115698", type="mass_ppm", browse=TRUE, mass_entry="5")
buildCompToxURL("162.115698", type="mass_ppm", browse=TRUE, mass_entry="5", single_component = FALSE,
                no_isotopes = FALSE)
buildCompToxURL("162.115698", type="mass_ppm", browse=TRUE, mass_entry="100")
buildCompToxURL("162.115698", type="mass_Da", browse=TRUE, mass_entry="0.01")
buildCompToxURL("162.115698", type="mass_Da", browse=TRUE, mass_entry="0.1")
buildCompToxURL("162.115698", type="mass_Da", browse=TRUE, mass_entry="0.1",
                single_component = FALSE, no_isotopes = FALSE)
# Search by mass range
buildCompToxURL("162.115", type="mass_range", browse=TRUE, mass_entry="162.116",
                single_component = FALSE, no_isotopes = FALSE)

}
\author{
Emma Schymanski (R wrapper, <emma.schymanski@uni.lu>), Antony Williams (CompTox Dashboard)
}

