% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildStructandSeries.R
\name{buildSeriesLabel}
\alias{buildSeriesLabel}
\title{Build Labels for Homologue Series}
\usage{
buildSeriesLabel(genSmiles, SeriesName, R1toN, nR1toN, ExtraAtoms_R1toN=NULL,
RDB_R1toN=NULL, split_nR1toN=FALSE)
}
\arguments{
\item{genSmiles}{A "Generic" \code{SMILES} code to expand, with R groups encoded in square brackets.}

\item{SeriesName}{A general name for the series, preferably short (abbreviation).}

\item{R1toN}{A vector containing 1 to n \code{SMILES} entries representing the repeating unit. One entry
per R group is required, comma separated and enclosed in brackets. Examples \code{"(C,C,C)", "(C,CCO)"}.}

\item{nR1toN}{The range for each R group (start-end,start-end, ...,start-end).
If a single range is given for n R groups, this can optionally be split by setting \code{split_nR1toN}.
Different labels are produced depending on this setting.}

\item{ExtraAtoms_R1toN}{(optional) If present, is added to the end of the label.}

\item{RDB_R1toN}{(optional) If present, is added to the end of the label (after ExtraAtoms)}

\item{split_nR1toN}{If \code{TRUE}, splits \code{nR1toN} with \code{\link{splitRrange}}.}
}
\value{
Returns a label that can be used to annotate generic structures in CDK Depict
}
\description{
This function generates labels to plot homologue series with the inputs used
in the \code{buildSmiles} function.
}
\examples{
genSmiles <- "OS(=O)(=O)c1ccc(cc1)C([R1])[R2]"
buildSeriesLabel(genSmiles,"LAS","{C,C}","{9-13}")
buildSeriesLabel(genSmiles,"LAS","{C,C}","{9-13}", split_nR1toN=TRUE)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{buildSmiles}}, \code{\link{splitRrange}}.
}

