% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildStructandSeries.R
\name{buildSmiles}
\alias{buildSmiles}
\title{Build SMILES from R groups}
\usage{
buildSmiles(genSmiles, R1toN, nR1toN, ExtraAtoms_R1toN=NULL, RDB_R1toN=NULL)
}
\arguments{
\item{genSmiles}{A "Generic" \code{SMILES} code to expand, with R groups encoded in square brackets.
The first group can be \code{"[R]"} or \code{"[R1]"}, subsequent groups should be numbered
systematically (e.g. \code{"[R2]", "[R3]", ...}). Examples \code{"O=S(O)(=O)OCCO[R1]C[R2]", "CCO[R]CCO"}}

\item{R1toN}{A vector containing 1 to n \code{SMILES} entries representing the repeating unit. One entry
per R group is required, comma separated and enclosed in brackets. Examples \code{"(C,C,C)", "(C,CCO)"}.}

\item{nR1toN}{The range for each R group (start-end,start-end, ...,start-end).
If a single range is given for n R groups, this should be split in advance with \code{splitRrange}}

\item{ExtraAtoms_R1toN}{(optional, to be removed from this function and handled in advance).
Extra atoms to be added to beginning or end of R group. Must be valid \code{SMILES} to generate
valid structures. Use with caution.}

\item{RDB_R1toN}{(optional, to be removed from this function and handled in advance).
Unsaturation to be added to beginning or end of R group. Must be valid \code{SMILES} to generate
valid structures. Use with caution - double-bond stereochemistry is not created.}
}
\value{
Returns a vector containing the resulting SMILES codes
}
\description{
This function performs basic string manipulation on SMILES codes to generate
systematic homologues, facilitaing the generation of identifiers for structures
that may not exist in databases for use in MS workflows.
This requires a decent understanding of SMILES as the text manipulation can have
unexpected consequences. This does not perform any advanced chemical processing.
Using large ranges over many R groups will result in combinatorial explosion; for
these cases structure generation is recommended. SMILES are built with the
\code{RtoSMILES} function.
}
\examples{
test_smiles <- "P(=O)(OC[C@H](COP(=O)(O)OC[C@H](OC(=O)[R2])COC(=O)[R1])O)(O)OC[C@H](OC(=O)[R4])COC(=O)[R3]"
buildSmiles(test_smiles,"(C,C,C,C)","(1-3,1-3,1-3,1-3)")

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{RtoSMILES}}, \code{\link{splitRrange}}, \code{\link{adjustRgroup}},
To view created SMILES: \code{\link{renderSMILES.CDKdepict}}, \code{\link{renderSMILES.rcdk}}.
}

