% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{CAS_fix}
\alias{CAS_fix}
\title{Fix CAS numbers if they fail CAS_test}
\usage{
CAS_fix(CAS_RN)
}
\arguments{
\item{CAS_RN}{Text string containing the CAS number to fix.}
}
\value{
Returns the fixed CAS number.
}
\description{
This function uses basic string manipulation to split
CAS numbers that have been converted to dates in excel. The CAS number
should have failed \code{\link{CAS_test}} prior to fixing.
}
\examples{
CAS_fix("95-14-7")
CAS_fix("01.03.7440")
CAS_fix("1/12/5392")
CAS_fix("NotCASNumber")

}
\seealso{
\code{\link{CAS_test}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
