% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{CAS_test}
\alias{CAS_test}
\title{Test whether CAS number follows valid format}
\usage{
CAS_test(CAS_RN)
}
\arguments{
\item{CAS_RN}{Text string containing the CAS number to test.}
}
\value{
Returns \code{TRUE} or \code{FALSE} (fails test).
}
\description{
This basic function using string manipulation to test whether a CAS number
is in the expected format. Useful to detect errors that occur in excel etc.
Potential future extension: to return corrected CAS numbers (not yet implemented).
This function checks if the string is longer than 7 and contains 3 dashes.
}
\examples{
CAS_test("95-14-7")
CAS_test("01.03.7440")

}
\seealso{
\code{\link{getSuspectIdentifiers}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
