% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{check_ded2}
\alias{check_ded2}
\title{Check if "deduct" can be subtracted from given molecular formula}
\usage{
check_ded2(formula, deduct)
}
\arguments{
\item{formula}{Molecular formula to check.}

\item{deduct}{Deduct (molecular formula to subtract from \code{formula}).}
}
\value{
Returns \code{TRUE} if the \code{deduct} can be subtracted from \code{formula}, otherwise
\code{FALSE}.
}
\description{
This function uses the formula functionality in \code{RMassBank} to check
if a formula can be subtracted from another (i.e. is a subset of another), especially for
calculating adduct masses. This serves in the place of \code{check_ded} from \code{enviPat},
which returns the opposite result.
}
\examples{
check_ded2("C6H6", "H2")
check_ded2("C6H6", "H3O")
check_ded2("C6H6","")

}
\seealso{
\code{\link{check_ded}}, \code{\link{enviPat}}, \code{\link{getSuspectFormulaMass}},
\code{\link{formulastring.to.list}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
