% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getCDKlogPs}
\alias{getCDKlogPs}
\title{Use rcdk to calculate AlogP and XlogP from SMILES}
\usage{
getCDKlogPs(smiles, kekulise=TRUE)
}
\arguments{
\item{smiles}{Valid SMILES code used to perform calculation.}

\item{kekulise}{Default \code{TRUE} performs aromaticity detection prior to calculation.
\code{FALSE} interprets the SMILES "as is". Can result in dramatically different logP values.}
}
\value{
Returns the resulting AlogP and XlogP values in a list.
}
\description{
A small wrapper function to calculate AlogP and XlogP
with rcdk from a SMILES code.
}
\examples{
getCDKlogPs("c1ccccc1")
getCDKlogPs("c1ccccc1",kekulise=FALSE)
getCDKlogPs("c1ccc2c(c1)[nH]nn2")
getCDKlogPs("c1ccc2c(c1)[nH]nn2",kekulise=FALSE)

}
\seealso{
\code{\link{parse.smiles}}, \code{\link{get.alogp}}, \code{\link{get.xlogp}},
\code{\link{getCDKlogPsAndRT}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
