% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getCDKlogPsAndRT}
\alias{getCDKlogPsAndRT}
\title{Use rcdk to calculate AlogP and XlogP from SMILES and estimate retention times}
\usage{
getCDKlogPsAndRT <- function(smiles,kekulise=TRUE,model=c(3.654,1.6225),quiet=TRUE)
}
\arguments{
\item{smiles}{Valid SMILES code used to perform calculation.}

\item{kekulise}{Default \code{TRUE} performs aromaticity detection prior to calculation.
\code{FALSE} interprets the SMILES "as is". Can result in dramatically different logP values.}

\item{model}{A vector of length 2 containing the intercept and slope for the model. Default values
\code{c(3.654,1.6225)} arise from the standard Eawag non-target chromatographic program on 810 standards.}
}
\value{
Returns the resulting AlogP, XlogP and RT estimated from XlogP values in a list.
Default model returns RT in minutes.
}
\description{
A small wrapper function to calculate AlogP and XlogP
with rcdk from a SMILES code (as \code{\link{getCDKlogPs}}) plus estimate retention
times using a simple linear model. Default model was calculated on 810 Eawag reference
standards and is only valid for this Xbridge chromatographic program.
}
\examples{
getCDKlogPsAndRT("c1ccccc1")
getCDKlogPsAndRT("c1ccccc1",kekulise=FALSE)
getCDKlogPsAndRT("c1ccc2c(c1)[nH]nn2")
getCDKlogPsAndRT("c1ccc2c(c1)[nH]nn2",kekulise=FALSE)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{parse.smiles}}, \code{\link{get.alogp}}, \code{\link{get.xlogp}},
\code{\link{getCDKlogPs}}.
}

