% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getInChI.obabel}
\alias{getInChI.obabel}
\title{Convert SMILES to an InChI with OpenBabel (obabel)}
\usage{
getInChI.obabel(smiles,babel_dir)
}
\arguments{
\item{smiles}{SMILES code to convert to the InChI.}

\item{babel_dir}{Location of folder containing \code{"obabel.exe"}.}
}
\value{
Returns the InChI retrieved, or alternative output from Babel.
}
\description{
A small wrapper function to convert SMILES to an InChI with OpenBabel
(\url{http://openbabel.org/wiki/Main_Page}). Requires pre-installation of
OpenBabel. 
This function uses Babel default InChI options; standard InChIs will be generated.
}
\examples{
babel_dir <- "C:/Program Files (x86)/OpenBabel-2.3.2"
getInChI.obabel("c1ccccc1", babel_dir)
Various failed conversions:
getInChI.obabel("blah", babel_dir)
getInChI.obabel("", babel_dir)

}
\seealso{
\code{\link{getInChI.obabel}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
