% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getInChIKey.obabel}
\alias{getInChIKey.obabel}
\title{Convert SMILES to an InChIKey with OpenBabel (obabel)}
\usage{
getInChIKey.obabel(smiles,babel_dir)
}
\arguments{
\item{smiles}{SMILES code to convert to the InChIKey.}

\item{babel_dir}{Location of folder containing \code{"obabel.exe"}.}
}
\value{
Returns the InChIKey retrieved, or alternative output from Babel. Run
\code{\link{InChIKey_test}} to determine if valid.
}
\description{
A small wrapper function to convert SMILES to an InChIKey with OpenBabel
(\code{\link{http://openbabel.org/wiki/Main_Page}}). Requires pre-installation of
OpenBabel. If this is not the case, use webservices instead via
\code{\link{getSuspectInChIKey}}. Note this supercedes \code{\link{getSuspectInChIKey.babel}}
This function uses Babel default InChI options; standard InChIKeys will be generated.
}
\examples{
babel_dir <- "C:/Program Files (x86)/OpenBabel-2.3.2"
getInChIKey.obabel("c1ccccc1", babel_dir)
Various failed conversions:
getInChIKey.obabel("blah", babel_dir)
getInChIKey.obabel("", babel_dir)
InChIKey_test(getInChIKey.obabel("blah",babel_dir))

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{getSuspectInChIKey}}, \code{\link{InChIKey_test}}, \code{\link{getSuspectInChIKey.babel}}.
}

