% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getPCdesc.title}
\alias{getPCdesc.title}
\title{Retrieve Record Title (Compound Name) from PubChem}
\usage{
getPCdesc.title(query, from = "cid", timeout=10)
}
\arguments{
\item{query}{string of the identifier (CID, InChIKey) to be converted}

\item{from}{Type of input ID (default \code{"cid"}, i.e. PubChem Compound ID, 
alternative \code{"inchikey"} to retrieve via InChIKey, or others (untried)).}

\item{timeout}{The timeout, in seconds.}
}
\value{
A list containing the CID and related record title (Name)
}
\description{
Retrieves record title (title name) from PubChem using PUG REST.
Default behaviour is to retrieve by CID or InChIKey.
}
\details{
For this function to work as expected, only one search entry should be
used (e.g. one CID or InChIKey). The URL can accept comma separated CIDs, 
but this is currently ignored downstream. 
Thanks to Paul Thiessen and Evan Bolton from PubChem team for assistance.
}
\examples{
getPCdesc.title("22206")
getPCdesc.title("FZXISNSWEXTPMF-UHFFFAOYSA-N",from="inchikey")
getPCdesc.title("2244")
getPCdesc.title("1983")
#a non-live record 
getPCdesc.title("4644")

}
\references{
PubChem search: \url{http://pubchem.ncbi.nlm.nih.gov/} 

PubChem PUG REST:
\url{https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest}
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
