% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getPCID.smiles}
\alias{getPCID.smiles}
\title{Retrieve PubChem CIDs from SMILES or InChI}
\usage{
getPCID.smiles(query, from = "smiles", to="cids")
}
\arguments{
\item{query}{SMILES or InChI (or other) to be converted}

\item{from}{Type of input ID (default \code{"smiles"}, i.e. SMILES, alternative
\code{"inchi"} to retrieve via InChI, or others (untried))}
}
\value{
The matching CID
}
\description{
Retrieves PubChem CIDs from SMILES or InChI using PUG REST.
By "default", this also returns the CID of the prefered tautomer
according to current PubChem standardization procedures
}
\details{
This searches PubChem by SMILES and/or InChI and returns CIDs. The function 
may work for other output, but this has not been tested.
Thanks to Paul Thiessen and Evan Bolton from PubChem team for assistance.
}
\examples{
getPCID.smiles("CC(=O)OC1=CC=CC=C1C(=O)O")

}
\references{
PubChem search: \url{http://pubchem.ncbi.nlm.nih.gov/} 

Pubchem REST:
\url{https://pubchem.ncbi.nlm.nih.gov/pug_rest/PUG_REST.html}
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
