% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getPCInChIKey}
\alias{getPCInChIKey}
\title{Retrieve InChIKey from PubChem CID}
\usage{
getPCInChIKey(query, from = "cid", to="InChIKey")
}
\arguments{
\item{query}{ID to be converted}

\item{from}{Type of input ID (default \code{cid}, i.e. PubChem CID)}
}
\value{
The InChIKey
}
\description{
Retrieves InChIKey from PubChem CID using PUG REST.
}
\details{
This searches PubChem for the CID and returns the InChIKey. The function 
may work for other output, but this has not been tested.
}
\examples{
getPCInChIKey("2519")

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\references{
PubChem search: \url{http://pubchem.ncbi.nlm.nih.gov/} 

Pubchem REST:
\url{https://pubchem.ncbi.nlm.nih.gov/pug_rest/PUG_REST.html}
}

