% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getSmilesFromInChI.babel}
\alias{getSmilesFromInChI.babel}
\title{Convert an InChI to SMILES with OpenBabel}
\usage{
getSmilesFromInChI.babel(inchi,babel_dir,temp_dir)
}
\arguments{
\item{babel_dir}{Location of folder containing \code{"babel.exe"}.}

\item{temp_dir}{Location of folder to save \code{"temp_inchi.inchi"}.}

\item{smiles}{Valid InChI code used to convert to SMILES. Default babel settings are used.}
}
\value{
Returns the SMILES retrieved, or alternative output from Babel.
}
\description{
A small wrapper function to convert an InChI to SMILES with OpenBabel
(\code{\link{http://openbabel.org/wiki/Main_Page}}). Requires pre-installation of
OpenBabel. If this is not the case, use webservices instead via
\code{\link{getCactus}}.  Note this creates \code{"temp_inchi.inchi"}
in the \code{temp_dir} directory and will overwrite any previous file of the same name.
}
\examples{
inchi <- "InChI=1S/C2H4O2/c1-2(3)4/h1H3,(H,3,4)"
getSmilesFromInChI.babel(inchi, "C:/Program Files (x86)/OpenBabel-2.3.2", "C:/DATA")

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{getCactus}}, \code{\link{getSuspectInChIKey.babel}}.
}

