% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getSuspectFormulaMass}
\alias{getSuspectFormulaMass}
\title{Calculate Molecular Formula and Exact Masses from SMILES}
\usage{
getSuspectFormulaMass(smiles)
}
\arguments{
\item{smiles}{Valid SMILES code used to calculate the formula and from then on the masses.}
}
\value{
Returns the molecular formula, monoisotopic mass, [M+H]+, [M+NH4]+, [M+Na]+ and [M-H]-
masses in a named list. If no H is present for [M-H]-, NA is returned.
}
\description{
A small wrapper function to calculate the molecular formula
from SMILES with the rcdk, as well as selected MS-relevant exact masses for fixed adducts
using enviPat. This uses \code{getMolecule} instead of \code{parse.smiles}.
Note: current version does not handle labelling as rcdk is used to obtain the formula.
See \code{\link{getSuspectMasses}} for an extension handling all enviPat adducts.
}
\examples{
getSuspectFormulaMass("c1ccccc1")


}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{getMolecule}}, \code{\link{get.mol2formula}}, \code{\link{enviPat}},
\code{\link{getSuspectMasses}}.
}

