% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getSuspectIdentifiers}
\alias{getSuspectIdentifiers}
\title{Get PubChem CID, ChemSpider ID and CAS from InChIKey via webservices}
\usage{
getSuspectIdentifiers(InChIKey, allCAS=FALSE, sep=",", minCAS=TRUE)
}
\arguments{
\item{InChIKey}{Valid InChIKey used to retrieve information.}

\item{allCAS}{Default \code{FALSE} will result in only one CAS number.
If \code{TRUE}, all CAS are returned in one string, separated by \code{sep}.
This overrides \code{minCAS}.}

\item{sep}{Determines the separator used with \code{allCAS}, default \code{","}.}

\item{minCAS}{Default \code{TRUE} returns the smallest CAS number (shortest string) if
\code{allCAS=FALSE}. If \code{FALSE}, returns the first CAS number in the list.}
}
\value{
Returns a list with the resulting identifiers, with \code{NA} if the query failed.
}
\description{
A wrapper function to obtain identifiers from an InChIKey
using various webservices. PubChem CIDs are obtained with \code{\link{getPcID}} from
PubChem direct (\url{https://pubchem.ncbi.nlm.nih.gov/}), ChemSpider IDs
directly from ChemSpider (\url{http://www.chemspider.com/}) with
\code{\link{getCSID}}. CAS numbers are retrieved using \code{\link{getCactus}} from
Cactus (\url{https://cactus.nci.nih.gov/chemical/structure});
if this fails the Chemical Translation Service (CTS,
\url{http://cts.fiehnlab.ucdavis.edu/} is queried. Options are given to
return all CAS numbers, the "smallest" CAS number or the first entry.
}
\examples{
benz_key <- "UHOVQNZJYSORNB-UHFFFAOYSA-N"
getSuspectIdentifiers(benz_key)
getSuspectIdentifiers(benz_key,allCAS=TRUE)
getSuspectIdentifiers(benz_key,allCAS=FALSE,minCAS=TRUE)
getSuspectIdentifiers(benz_key,allCAS=FALSE,minCAS=FALSE)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{getCactus}}, \code{\link{getPcId}}, \code{\link{getCSID}}
, \code{\link{getCtsRecord}}, \code{\link{getSuspectInChIKey}}.
}

