% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getSuspectInChIKey.babel}
\alias{getSuspectInChIKey.babel}
\title{Convert SMILES to an InChIKey with OpenBabel (superceded)}
\usage{
getSuspectInChIKey.babel(smiles,babel_dir,temp_dir)
}
\arguments{
\item{smiles}{Valid SMILES code used to convert to the InChIKey.}

\item{babel_dir}{Location of folder containing \code{"babel.exe"}.}

\item{temp_dir}{Location of folder to save \code{"temp_smiles.smi"}.}
}
\value{
Returns the InChIKey retrieved, or alternative output from Babel. Run
\code{\link{InChIKey_test}} to determine if valid.
}
\description{
A small wrapper function to convert SMILES to an InChIKey with OpenBabel
(\url{http://openbabel.org/wiki/Main_Page}). Requires pre-installation of
OpenBabel. If this is not the case, use webservices instead via
\code{\link{getSuspectInChIKey}}. Note this creates \code{"temp_smiles.smi"}
in the \code{temp_dir} directory and will overwrite any previous file of the same name.
This function uses Babel default InChI options; standard InChIKeys will be generated.
This is superceded by \code{\link{getInChIKey.obabel}}.
}
\examples{
getSuspectInChIKey.babel("c1ccccc1", "C:/Program Files (x86)/OpenBabel-2.3.2", "C:/DATA")
Various failed conversions:
getSuspectInChIKey.babel("blah", "C:/Program Files (x86)/OpenBabel-2.3.2", "C:/DATA")
getSuspectInChIKey.babel("", "C:/Program Files (x86)/OpenBabel-2.3.2", "C:/DATA")

}
\seealso{
\code{\link{getSuspectInChIKey}}, \code{\link{getInChIKey.obabel}}, \code{\link{InChIKey_test}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
