% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getSuspectInChIKey}
\alias{getSuspectInChIKey}
\title{Retrieve an InChIKey by SMILES or Name with Cactus/CTS}
\usage{
getSuspectInChIKey(smiles,name)
}
\arguments{
\item{smiles}{Valid SMILES code used to retrieve the InChIKey.}

\item{name}{Chemical name used to query CTS. Strange names will likely fail.}
}
\value{
Returns the InChIKey retrieved, or \code{NA}.
}
\description{
A small wrapper function to retrieve an InChIKey with webservices. Cactus
(\url{https://cactus.nci.nih.gov/chemical/structure}) is used to obtain a Key
from SMILES; if this fails the Chemical Translation Service (CTS,
\url{http://cts.fiehnlab.ucdavis.edu/} is queried by name.
Future extensions possible. If SMILES is available and Open Babel is installed,
use \code{\link{getSuspectInChIKey.babel}} instead.
}
\examples{
getSuspectInChIKey("c1ccccc1", "benzene")

}
\seealso{
\code{\link{getSuspectInChIKey.babel}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
