% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{getSuspectMasses}
\alias{getSuspectMasses}
\title{Calculate Masses for various adducts from SMILES}
\usage{
getSuspectMasses(smiles, adduct_list)
}
\arguments{
\item{smiles}{Valid SMILES code used to calculate the formula and from then on the masses.}

\item{adduct_list}{A list of adducts matching those in the \code{enviPat} \code{adducts} table.}
}
\value{
Returns the masses of the given adducts in a matrix with column names matching the given adduct.
If the adduct/deduct cannot be calculated for a given entry, NA is returned.
}
\description{
A small wrapper function to calculate adduct masses for any entries in
the \code{enviPat} \code{adducts} table from the SMILES code, using
the \code{getMolecule} function in \code{rcdk} to calculate the formula from SMILES first.
Note: current version does not handle labelling as rcdk is used to obtain the formula.
}
\examples{
getSuspectMasses("c1ccccc1", c("M+H","M+NH4"))
getSuspectMasses("c1ccccc1", c("M-H","M-H2O-H"))
getSuspectMasses("c1ccccc1", c("M+H","M+2H","M+3Na"))


}
\seealso{
\code{\link{getMolecule}}, \code{\link{get.mol2formula}}, \code{\link{enviPat}},
\code{\link{check_ded2}}, \code{\link{getSuspectFormulaMass}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
