% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{InChIKey_test}
\alias{InChIKey_test}
\title{Test whether InChIKey follows valid format}
\usage{
InChIKey_test(InChIKey)
}
\arguments{
\item{InChIKey}{InChIKey to test.}
}
\value{
Returns \code{TRUE} or \code{FALSE} (fails test).
}
\description{
This basic function using string manipulation to test whether an InChIKey
is in the expected format. Useful to detect conversions and retrieval results.
This function checks if the string contains 27 characters and 2 dashes/3 blocks.
}
\examples{
CactusKey <- getCactus("benzene","stdinchikey")
InChIKey_test(CactusKey)
CactusKey <- trimKey(CactusKey)
InChIKey_test(CactusKey)
InChIKey_test("UHOVQNZJYSORNB-UHFFFAOYSA-N")

}
\seealso{
\code{\link{getSuspectInChIKey}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
