% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{MolFormFromInChI}
\alias{MolFormFromInChI}
\title{Extract Molecular Formula from InChI}
\usage{
MolFormFromInChI(InChI)
}
\arguments{
\item{InChI}{InChI string to extract the formula.}
}
\value{
Returns the molecular formula as is in the InChI string.
}
\description{
This basic function using string manipulation to extract a formula from
the front of an InChI string, returning the second field (the formula).
Note: this does NOT account for labeling, i.e. it is the formula associated
with the structural skeleton (the InChIKey first block).
}
\examples{
inchi <- "InChI=1S/C2H4O2/c1-2(3)4/h1H3,(H,3,4)"
MolFormFromInChI(inchi)

}
\seealso{
\code{\link{MolFormFromSmiles.rcdk}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
