% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{MolFormFromSmiles.rcdk}
\alias{MolFormFromSmiles.rcdk}
\title{Calculate Molecular Formula from SMILES with rcdk}
\usage{
MolFormFromSmiles.rcdk(smiles)
}
\arguments{
\item{smiles}{Valid SMILES code used to calculate the formula.}
}
\value{
Returns the molecular formula only (the \code{rcdk} function contains additional text).
}
\description{
A small wrapper function to calculate the molecular formula
from SMILES with the rcdk, as recommended. Note: current version does not handle labelling.
}
\examples{
MolFormFromSmiles.rcdk("OC(=O)C")

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{parse.smiles}}, \code{\link{get.mol2formula}}, \code{\link{MolFormFromInChI}}.
}

