% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructurePlotting.R
\name{renderSMILES.CDKdepict}
\alias{renderSMILES.CDKdepict}
\title{Render SMILES into 2D image for plotting via CDK Depict}
\usage{
renderSMILES.CDKdepict(depictURL,coords=c(0,0,100,100), filename="tmp.svg")
}
\arguments{
\item{depictURL}{A URL created with \code{buildCDKdepictURL}}

\item{coords}{This is used to control the size of the image within the plot. Values
\code{c(xmin,ymin,xmax,ymax)}.}

\item{filename}{Name of the temporary svg file for saving the image. Can be used
to save all image files if renamed for every instance.}
}
\value{
Returns an image for use during plotting
}
\description{
This function uses the URL built with \code{buildCDKdepictURL} to create an
image for direct use in plotting, via a temporary svg file and given
coordinates. Output requires that plot.new has been called.
}
\examples{
plot.new()
plot.window(xlim=c(0,200), ylim=c(0,100))
test_url <- buildCDKdepictURL("OS(=O)(=O)c1ccc(cc1)C(CC(=O)O)CC(=O)O |Sg:n:15:m:ht,Sg:n:11:n:ht|",
style="cow",title="SPADCs, n+m=0-5", abbr="off",suppressh=TRUE,showtitle=TRUE,zoom=5,annotate="off",
smarts="C(=O)O")
renderSMILES.CDKdepict(test_url)

}
\seealso{
\code{\link{buildCDKdepictURL}} to create the URL, \code{\link{renderSMILES.rcdk}}
for alternative plotting direct from SMILES.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
