% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructurePlotting.R
\name{renderSMILES.rcdk.default}
\alias{renderSMILES.rcdk.default}
\title{Render SMILES into 2D image for plotting via rcdk}
\usage{
renderSMILES.rcdk.default(smiles, kekulise=TRUE, coords=c(0,0,100,100))
}
\arguments{
\item{smiles}{A valid SMILES code for rendering (e.g. \code{"c1ccccc1"}).}

\item{kekulise}{If \code{TRUE}, performs CDK aromaticiy detection, which is recommended.
Setting \code{FALSE} can force rendering of invalid SMILES with undefined bond orders.
Older rcdk versions may behave differently.}

\item{coords}{This is used to control the size of the image within the plot. Values
\code{c(xmin,ymin,xmax,ymax)}.}
}
\value{
Returns an image for use during plotting
}
\description{
This function uses the rcdk to parse the smiles into a mol, with options
to switch kekulise (aromaticity detection) on or off and to define desired
coordinates. Output requires that plot.new has been called, i.e. this is
designed to be used directly during plotting.
This function uses default depiction options. For more options with
rcdk>3.4.1, use \code{\link{renderSMILES.rcdk}}. Best results with rcdk>3.4.1 and
rcdklibs>2.0.
}
\details{
More information about aromaticity: \code{\link{https://github.com/rajarshi/cdkr/issues/49}}
}
\examples{
smiles <- "OS(=O)(=O)c1ccc(cc1)C(CC(=O)O)CC(=O)O"
  plot.new()
  plot.window(xlim=c(0,200), ylim=c(0,100))
  renderSMILES.rcdk(smiles,kekulise=FALSE)
  renderSMILES.rcdk(smiles,kekulise=TRUE)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{renderSMILES.CDKdepict}}, \code{\link{renderSMILES.rcdk}}
}

