% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructurePlotting.R
\name{renderSMILES.rcdk}
\alias{renderSMILES.rcdk}
\title{Render SMILES into 2D image for plotting via rcdk}
\usage{
renderSMILES.rcdk(smiles, kekulise=TRUE, coords=c(0,0,100,100),
width=200, height=200, zoom=1.3,style="cow", annotate="off", abbr="on",suppressh=TRUE,
showTitle=FALSE, smaLimit=100, sma=NULL)
}
\arguments{
\item{smiles}{A valid SMILES code for rendering (e.g. \code{"c1ccccc1"}).}

\item{kekulise}{If \code{TRUE}, performs CDK aromaticiy detection, which can lead to
undesirable plotting results with older versions. Setting \code{FALSE} renders
the SMILES code as is but will plot aromatic structures with undefined bond orders.}

\item{coords}{This is used to control the size of the image within the plot. Values
\code{c(xmin,ymin,xmax,ymax)}.}

\item{width}{Defines width of depiction object}

\item{height}{Defines height of depiction object}

\item{zoom}{Controls the size of the image, default 2, increasing to 5 results in a large
structure in the web browser. Warnings at zoom>=10.}

\item{style}{The plotting style, one of 5 options \code{("cow","cob","bow","wob","nob")},
i.e. color on white, color on black, black on white, white on black, neon on black. The
appearance of \code{SMARTS} and code{annotate} options change with style.}

\item{annotate}{(optional) Features to interpret the structure. Default \code{"none"}, other
options Atom Numbers, Atom Mapping, Color Map, Atom Value
(\code{"number", "mapidx", "colmap", "atomvalue"}) respectively.}

\item{abbr}{Default \code{"off"}, this controls whether the structure is displayed "as is"
\code{"off"} or whether groups (\code{"groups"}), reagents (\code{"reagents"}) or both reagents
and groups (\code{"on"}) are abbreviated.}

\item{suppressh}{Default \code{TRUE} suppresses the display of Hs. \code{FALSE} does the oppposite.}

\item{showTitle}{Default \code{FALSE} turns title display off (no title is included)}

\item{smaLimit}{sets the limit for \code{SMARTS} patterns to highlight}

\item{sma}{(optional) \code{SMARTS} codes can be entered to highlight parts of the molecule.}
}
\value{
Returns an image for use during plotting
}
\description{
This function uses the rcdk to parse the smiles into a mol, with options
to switch aromaticity detection on or off and to define desired
coordinates, as well as depiction control. Compatible only with rdck above 3.4.1. Default
parameters produce the same as \code{\link{renderSMILES.rcdk.default}}.
Output requires that plot.new has been called, i.e. this is
designed to be used directly during plotting.
Users with rcdk<3.4.1 should use \code{\link{renderSMILES.rcdk.default}} or
\code{renderSMILES.CDKdepict} to view latest CDK updates.
As opposed to older versions, kekulise=FALSE now produces strange behaviour for some
SMILES.
}
\details{
More information about aromaticity: \url{https://github.com/rajarshi/cdkr/issues/49}
}
\examples{
smiles <- "OS(=O)(=O)c1ccc(cc1)C(CC(=O)O)CC(=O)O"
plot.new()
plot.window(xlim=c(0,200), ylim=c(0,100))
renderSMILES.rcdk(smiles,kekulise=FALSE)
renderSMILES.rcdk(smiles,kekulise=TRUE)
renderSMILES.rcdk(smiles,kekulise=TRUE, abbr="off")

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{view.molecule.2d}} \code{\link{renderSMILES.CDKdepict}}
}

