% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitPFAS.R
\name{splitPFAS.csv}
\alias{splitPFAS.csv}
\title{Run SplitPFAS on a CSV file containing SMILES}
\usage{
splitPFAS.csv(
  pfas_file,
  splitPFAStools_name,
  smarts_file,
  babel_dir,
  incl_image = TRUE,
  smiles_col_index = 3,
  SMILES_KeyCheck = TRUE,
  func_group_col_index = 5,
  Func_Group_KeyCheck = FALSE,
  pfas_part_col_index = 6,
  PFAS_Part_KeyCheck = FALSE
)
}
\arguments{
\item{pfas_file}{Name of the CSV file containing at least one SMILES column}

\item{splitPFAStools_name}{Name of the SplitPFAS jar. See \code{\link{SplitPFAS}}}

\item{smarts_file}{Name of the SMARTS file. See \code{\link{SplitPFAS}}}

\item{babel_dir}{Location of Open Babel dir. Needed for InChIKey checks.}

\item{incl_image}{Whether to include image generation. See \code{\link{SplitPFAS}}}

\item{smiles_col_index}{Index of the column containing SMILES. Default 3.}

\item{SMILES_KeyCheck}{Boolean to indicate if SMILES InChIKey check should be performed.}

\item{func_group_col_index}{Index of column containing functional group. For validation only.}

\item{Func_Group_KeyCheck}{Default \code{FALSE}. Checks functional group InChIKeys if \code{TRUE} 
- only for test files that contain this information (advanced users only)}

\item{pfas_part_col_index}{Index of column containing functional group. For validation only.}

\item{PFAS_Part_KeyCheck}{Default \code{FALSE}. Checks functional group InChIKeys if \code{TRUE} 
- only for test files that contain this information (advanced users only).}
}
\description{
A little wrapper function to run SplitPFAS on a csv file
containing SMILES and other fields, optionally running InChIKey checks
NOTE: some options only work for very specific files. Use with caution.
}
\details{
Returns a csv file containing the SplitPFAS results.
}
\examples{
pfas_file <- system.file("extdata","PFAS_SplitSmiles_ExamplesExt.csv",package="RChemMass")
splitPFAStools_name <- "MetFrag2.4.5-Tools.jar"
smarts_file <- system.file("extdata","SplitPFAS_smarts.txt",package="RChemMass")
babel_dir <- "C:/Program Files/OpenBabel-2.4.1"
splitPFAS.csv(pfas_file, splitPFAStools_name, smarts_file, babel_dir,
              incl_image = TRUE, smiles_col_index = 3, SMILES_KeyCheck = TRUE,
              func_group_col_index = 5, Func_Group_KeyCheck = FALSE,
              pfas_part_col_index = 6, PFAS_Part_KeyCheck = FALSE)
              
splitPFAS.csv(pfas_file, splitPFAStools_name, smarts_file, babel_dir,
              incl_image = TRUE, smiles_col_index = 3, SMILES_KeyCheck = TRUE,
              func_group_col_index = 5, Func_Group_KeyCheck = TRUE, 
              pfas_part_col_index = 6, PFAS_Part_KeyCheck = TRUE)
              
pfas_file <- system.file("extdata","Fluroseedset_1to30.csv",package="RChemMass")
splitPFAS.csv(pfas_file, splitPFAStools_name, smarts_file, babel_dir,
              incl_image = TRUE, smiles_col_index = 5, SMILES_KeyCheck = TRUE,
              Func_Group_KeyCheck = FALSE, PFAS_Part_KeyCheck = FALSE)
              
}
\seealso{
\code{\link{splitPFAS.help}}, \code{\link{splitPFAS}}
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
