% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitPFAS.R
\name{splitPFAS.help}
\alias{splitPFAS.help}
\title{Display Help for SplitPFAS Jar}
\usage{
splitPFAS.help(splitPFAStools_name, splitPFAStools_dir="")
}
\arguments{
\item{splitPFAStools_name}{The name of the SplitPFAS Jar file (e.g. 
\code{"MetFrag2.4.5-Tools.jar"}).}

\item{splitPFAStools_dir}{The directory containing the jar file; if empty assumes the
currect directory.}
}
\value{
Returns the help text in the console
}
\description{
A small wrapper function to display the "help" text for SplitPFAS Jar.
The jar file is available from 
\url{https://msbi.ipb-halle.de/~cruttkie/metfrag/MetFrag2.4.5-Tools.jar}
}
\examples{
splitPFAStools_name <- "MetFrag2.4.5-Tools.jar"
splitPFAS.help(splitPFAStools_name)

}
\seealso{
\code{\link{splitPFAS}}
}
\author{
Emma Schymanski <emma.schymanski@uni.lu> in partnership with Christoph Ruttkies (MetFragTools)
and Zhanyun Wang (ETHZ, PFAS examples).
}
