% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitPFAS.R
\name{splitPFAS}
\alias{splitPFAS}
\title{Split Chemical (as SMILES) into Perfluorinated Part and Functional Group using SplitPFAS}
\usage{
splitPFAS(smiles, splitPFAStools_name, incl_image=FALSE, smarts_file="", 
splitPFAStools_dir="")
}
\arguments{
\item{smiles}{The SMILES code of the PFAS to split. Preferably an "MS-ready" form (no salts etc)}

\item{splitPFAStools_name}{The name of the SplitPFAS Jar file (e.g. 
\code{"MetFrag2.4.5-Tools.jar"}).}

\item{incl_image}{If \code{TRUE}, a PNG image showing the bond broken is saved in a subdirectory 
\code{"SplitPFAS_PNGs"} in the current directory. The jar creates a file in the temp directory 
which is copied over and deleted from the temp dir. Default \code{FALSE} switches off image creation.}

\item{smarts_file}{If empty, PFAS are split at the end of the perfluorinated chain. If a path
is given, the SMILES is split according to the matching SMARTS in the file, where order of the 
SMARTS in the file indicates priority matching order (first match is processed)}

\item{splitPFAStools_dir}{The directory containing the jar file; if empty assumes the
currect directory.}
}
\value{
Returns a list containing the processed SMILES (\code{SMILES_in}), the (non-PFAS) 
functional group (\code{Func_Group}), the PFAS Group(s) split off (\code{PFAS_Groups}), 
both of which are pipe-separated (\code{"|"}) if different groups are present; the number of PFAS 
groups split off (\code{n_PFAS_Groups}), the SMARTS string used to do the splitting 
\code{SplitSMARTS}, the PNG location (\code{PNG_location}), which is empty if PNG creation is off
and finally \code{Error_msg}, which contains an error message if issues were encountered.
}
\description{
A small wrapper function to run the SplitPFAS Jar to split perfluorinated compounds
(PFAS) into PFAS chain and functional group, by given SMARTS group(s). NOTE this is under 
development: Please contact the authors for more information. The jar is available from 
\url{https://msbi.ipb-halle.de/~cruttkie/metfrag/MetFrag2.4.5-Tools.jar}
}
\examples{
splitPFAStools_name <- "MetFrag2.4.5-Tools.jar"
smarts_file <- system.file("extdata","SplitPFAS_smarts.txt",package="RChemMass")
pfas_smiles <- "CN(CCOC(=O)C(C)=C)S(=O)(=O)C(F)(F)C(F)(F)C(F)(F)C(F)(F)C(F)(F)F"
pfas_smiles <- "FC(F)(F)C(F)(F)C(F)(F)C(F)(F)S(=O)(=O)CS(=O)(=O)C(F)(F)C(F)(F)C(F)(F)C(F)(F)F"
pfas_smiles <- "FC(F)(F)C(F)(F)C(F)(F)C(F)(F)C(F)(F)C(F)(F)CCOP(O)(OCCC(F)(F)C(F)(F)C(F)(F)C(F)(F)C(F)(F)C(F)(F)F)OCCC(F)(F)C(F)(F)C(F)(F)C(F)(F)C(F)(F)C(F)(F)F"
pfas_smiles <- "FCC(F)(F)C(F)C(F)(F)F"
pfas_smiles <- "CCC(CN)C(F)(F)C(F)(F)C(F)(F)F"

splitPFAS(pfas_smiles,splitPFAStools_name)
splitPFAS(pfas_smiles,splitPFAStools_name,smarts_file = smarts_file)
splitPFAS(pfas_smiles,splitPFAStools_name,smarts_file = smarts_file,incl_image = TRUE)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu> in partnership with Christoph Ruttkies (MetFragTools)
and Zhanyun Wang (ETHZ, PFAS examples).
}
\seealso{
\code{\link{splitPFAS.help}}
}

