% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildStructandSeries.R
\name{splitRrange}
\alias{splitRrange}
\title{Split a single R group range into n ranges}
\usage{
splitRrange(genSmiles, nR1toN)
}
\arguments{
\item{genSmiles}{A "Generic" \code{SMILES} code to expand, with R groups encoded in square brackets.
The first group can be \code{"[R]"} or \code{"[R1]"}, subsequent groups should be numbered
systematically (e.g. \code{"[R2]", "[R3]", ...}). Examples \code{"O=S(O)(=O)OCCO[R1]C[R2]", "CCO[R]CCO"}}

\item{nR1toN}{The range for each R group (start-end,start-end, ...,start-end) or a single range.
If a single range is given for n R groups, this is split to form a range for each R group.
Warnings are produced if mismatches occur, in which case the input is returned.}
}
\value{
Returns the split range or original input for use in \code{\link{buildSmiles}}
}
\description{
This splits a range into n for the case where e.g. the total number of carbons is
known but multiple Rs exist. This assumes symmetrical filling, with remainders
added to the earlier groups. Used to feed into the \code{buildSmiles} function
This function may perform strangely and needs testing.
}
\examples{
test_smiles <- "P(=O)(OC[C@H](COP(=O)(O)OC[C@H](OC(=O)[R2])COC(=O)[R1])O)(O)OC[C@H](OC(=O)[R4])COC(=O)[R3]"
splitRrange(test_smiles,"(1-37)")
splitRrange(test_smiles,"(1-38)")
splitRrange(test_smiles,"(1-39)")
splitRrange(test_smiles,"(1-40)")
splitRrange(test_smiles,"(3-40)")
splitRrange(test_smiles,"(1-3)")
splitRrange(test_smiles,"(16-22)")

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{RtoSMILES}}, \code{\link{buildSmiles}}.
}

