% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildStructandSeries.R
\name{trimBuiltSmiles}
\alias{trimBuiltSmiles}
\title{Trim SMILES list from buildSmiles to one SMILES per formula}
\usage{
trimBuiltSmiles(smiles_list)
}
\arguments{
\item{smiles_list}{A list of \code{SMILES} from \code{\link{buildSmiles}}.}
}
\value{
Returns a vector containing the resulting SMILES codes
}
\description{
This function takes the output of \code{\link{buildSmiles}} and
returns the first SMILES entry per molecular formula.
}
\examples{
LAS_gen_smiles <- "OS(=O)(=O)c1ccc(cc1)C([R1])[R2]"
LAS_smiles <- buildSmiles(LAS_gen_smiles, "{C,C}", splitRrange(LAS_gen_smiles, "{9-13}"))
trimBuiltSmiles(LAS_smiles)

}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
\seealso{
\code{\link{buildSmiles}}, \code{\link{MolFormFromSmiles.rcdk}}.
To view created SMILES: \code{\link{renderSMILES.CDKdepict}}, \code{\link{renderSMILES.rcdk}}.
}

