% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalCuration.R
\name{trimKey}
\alias{trimKey}
\title{Trim text from start of Cactus InChIKey}
\usage{
trimKey(InChIKey)
}
\arguments{
\item{InChIKey}{InChIKey to trim (or not).}
}
\value{
Returns the trimmed key or the original input.
}
\description{
This basic function trims \code{"InChIKey="} from the start of
InChIKeys obtained via Cactus with \code{\link{getCactus}}. Returns the original
Key if the preceding text is not present.
}
\examples{
CactusKey <- getCactus("benzene","stdinchikey")
trimKey(CactusKey)
trimKey("UHOVQNZJYSORNB-UHFFFAOYSA-N")

}
\seealso{
\code{\link{getCactus}}.
}
\author{
Emma Schymanski <emma.schymanski@uni.lu>
}
